Refactor vitals endpoints and functions into their own file
- Helps make index.js more readable
- Improves stsructure
- Makes it easier to write tests
