+++ server/routes.js
@@ -1,5 +1,6 @@
 const dashboard = require('./dashboard')
 const siren = require('./siren')
+const vitals = require('./vitals')
 
 function configureRoutes(app) {
   app.get('/', dashboard.sessionChecker, dashboard.redirectToHomePage)
@@ -20,6 +21,8 @@ function configureRoutes(app) {
   app.post('/locations/:locationId', dashboard.validateEditLocation, dashboard.submitEditLocation)
   app.post('/login', dashboard.submitLogin)
 
+  app.post('/api/devicevitals', vitals.validateDeviceVitals, vitals.handleDeviceVitals)
+  app.post('/api/heartbeat', vitals.validateHeartbeat, vitals.handleHeartbeat)
   app.post('/api/sirenAddressed', siren.validateSirenAddressed, siren.handleSirenAddressed)
   app.post('/api/sirenEscalated', siren.validateSirenEscalated, siren.handleSirenEscalated)
