+++ server/test/integration/serverTest.js
@@ -127,30 +127,6 @@ async function im21Door(doorCoreId, signal) {
   }
 }
 
-async function normalHeartbeat(coreId) {
-  try {
-    await chai.request(server).post('/api/heartbeat').send({
-      coreid: coreId,
-      data: `{"doorMissedMsg": 0, "doorLowBatt": false, "doorLastHeartbeat": 1000, "resetReason": "NONE", "states":[]}`,
-    })
-    await helpers.sleep(50)
-  } catch (e) {
-    helpers.log(e)
-  }
-}
-
-async function lowBatteryHeartbeat(coreId) {
-  try {
-    await chai.request(server).post('/api/heartbeat').send({
-      coreid: coreId,
-      data: `{"doorMissedMsg": 0, "doorLowBatt": true, "doorLastHeartbeat": 1000, "resetReason": "NONE", "states":[]}`,
-    })
-    await helpers.sleep(50)
-  } catch (e) {
-    helpers.log(e)
-  }
-}
-
 describe('Brave Sensor server', () => {
   beforeEach(async () => {
     await redis.clearKeys()
@@ -825,136 +801,6 @@ describe('Brave Sensor server', () => {
         const response = await chai.request(server).post('/api/devicevitals').send(badRequest)
         expect(response).to.have.status(400)
       })
-    })
-  })
-
-  describe('POST request heartbeat events with mock INS Firmware State Machine when battery level is normal', () => {
-    beforeEach(async () => {
-      await redis.clearKeys()
-      await db.clearTables()
-
-      const client = await clientFactory(db)
-      await locationFactory(db, {
-        locationid: testLocation1Id,
-        radarCoreId: radar_coreID,
-        doorCoreId: radar_coreID,
-        firmwareStateMachine: true,
-        clientId: client.id,
-      })
-
-      sandbox.stub(braveAlerter, 'startAlertSession')
-      sandbox.stub(braveAlerter, 'sendSingleAlert')
-      sandbox.spy(helpers, 'logSentry')
-      sandbox.stub(redis, 'addEdgeDeviceHeartbeat')
-    })
-
-    afterEach(async () => {
-      await redis.clearKeys()
-      await db.clearTables()
-      sandbox.restore()
-    })
-
-    it('should return 200 for a heartbeat request with no headers', async () => {
-      const response = await chai.request(server).post('/api/heartbeat').send({})
-      expect(response).to.have.status(200)
-    })
-
-    it('should return 200 for a valid heartbeat request', async () => {
-      const request = {
-        coreid: radar_coreID,
-        data: `{"doorMissedMsg": 0, "doorLowBatt": true, "doorLastHeartbeat": 1000, "resetReason": "NONE", "states":[]}`,
-      }
-      const response = await chai.request(server).post('/api/heartbeat').send(request)
-      expect(response).to.have.status(200)
-    })
-
-    it('should call addEdgeDeviceHeartbeat', async () => {
-      await normalHeartbeat(radar_coreID)
-      expect(redis.addEdgeDeviceHeartbeat).to.be.calledOnce
-    })
-
-    it('should not call logSentry', async () => {
-      await normalHeartbeat(radar_coreID)
-      expect(helpers.logSentry).to.not.be.called
-    })
-
-    it('should not update sentLowBatteryAlertAt if battery is normal', async () => {
-      const testLocation = await db.getLocationData(testLocation1Id)
-      await normalHeartbeat(radar_coreID)
-      const locationAfterHeartbeat = await db.getLocationData(testLocation1Id)
-      expect(locationAfterHeartbeat.sentLowBatteryAlertAt).to.deep.equal(testLocation.sentLowBatteryAlertAt)
-    })
-
-    it('should not call sendSingleAlert if battery level is normal', async () => {
-      await normalHeartbeat(radar_coreID)
-      expect(braveAlerter.sendSingleAlert).to.not.be.called
-    })
-  })
-
-  describe('POST request heartbeat events with mock INS Firmware State Machine when battery level is low', () => {
-    beforeEach(async () => {
-      await redis.clearKeys()
-      await db.clearTables()
-
-      const client = await clientFactory(db)
-      await locationFactory(db, {
-        locationid: testLocation1Id,
-        radarCoreId: radar_coreID,
-        doorCoreId: radar_coreID,
-        firmwareStateMachine: true,
-        sentLowBatteryAlertAt: firstLowBatteryAlert,
-        clientId: client.id,
-      })
-
-      sandbox.stub(braveAlerter, 'startAlertSession')
-      sandbox.stub(braveAlerter, 'sendSingleAlert')
-      sandbox.spy(helpers, 'logSentry')
-      sandbox.stub(redis, 'addEdgeDeviceHeartbeat')
-    })
-
-    afterEach(async () => {
-      await redis.clearKeys()
-      await db.clearTables()
-      sandbox.restore()
-    })
-
-    it('should return 200 for a heartbeat request with no headers', async () => {
-      const response = await chai.request(server).post('/api/heartbeat').send({})
-      expect(response).to.have.status(200)
-    })
-
-    it('should return 200 for a valid heartbeat request', async () => {
-      const request = {
-        coreid: radar_coreID,
-        data: `{"doorMissedMsg": 0, "doorLowBatt": true, "doorLastHeartbeat": 1000, "resetReason": "NONE", "states":[]}`,
-      }
-      const response = await chai.request(server).post('/api/heartbeat').send(request)
-      expect(response).to.have.status(200)
-    })
-
-    it('should call addEdgeDeviceHeartbeat', async () => {
-      await normalHeartbeat(radar_coreID)
-      expect(redis.addEdgeDeviceHeartbeat).to.be.called
-    })
-
-    it('should update sentLowBatteryAlertAt if the period since the last alert is greater than the timeout', async () => {
-      const locationBeforeHeartbeat = await db.getLocationData(testLocation1Id)
-      await lowBatteryHeartbeat(radar_coreID)
-      const locationAfterHeartbeat = await db.getLocationData(testLocation1Id)
-      expect(locationBeforeHeartbeat.sentLowBatteryAlertAt).to.not.deep.equal(locationAfterHeartbeat.sentLowBatteryAlertAt)
-    })
-
-    it('should not update sentLowBatteryAlertAt if the period since the last alert is less than the timeout', async () => {
-      await lowBatteryHeartbeat(radar_coreID)
-      const locationAfterHeartbeat = await db.getLocationData(testLocation1Id)
-      await lowBatteryHeartbeat(radar_coreID)
-      const locationAfterTwoHeartbeats = await db.getLocationData(testLocation1Id)
-      expect(locationAfterTwoHeartbeats.sentLowBatteryAlertAt).to.deep.equal(locationAfterHeartbeat.sentLowBatteryAlertAt)
-    })
-
-    it('should call sendSingleAlert', async () => {
-      await lowBatteryHeartbeat(radar_coreID)
-      expect(braveAlerter.sendSingleAlert).to.be.called
     })
   })
 })
